backend/pyproject.toml
[project]
name = "seo-dashboard-backend"
version = "0.1.0"
description = "SEO Dashboard API backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "weasyprint>=62.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
